Remove core Read() method and change Get() return behavior.
The core files.Get() method now returns None if a file doesn't exist.
Also, when called with multiple paths, Get() returns a dictionary mapping
paths to files.File objects. Only files that exist are present in the return.
This is a backwards-incompatible change, but makes the Get() interface much
clearer going forward.

Read() still exists in the HTTP API and the TitanClient, but no longer
in the core API or any service hooks.