Performance and API improvements by decoupling from entity objects.
This is a backwards-incompatible change to the API which removes all of the
"file_ent" args to core functions, and adds support for passing File objects
directly to the "path(s)" arg. Plugins and consumers can avoid extra RPCs by
passing fully-populated File objects, which removes the need to split this
information to paths= and file_ents= and also removes all high-coupling to
File._file_ent (at least outside of the files modules).

In the end, this will make this kind of operation perform about half the
number of RPCs, as well as significantly clean up the API and callers:

file_objs = files.Get(files.ListFiles('/some/dir', recursive=True)).values()
files.Delete(file_objs)